Make git and npm tags branch-aware
Make git and NPM semantic version tags branch aware. This means the
tags include the branch, if not the default, and increment
appropriately. Make publish parameter a choice, allowing publishing
only for default branch.  This also bases the version on the version
in the package.json, falling back to the next tag.

Move helper functions to separate files.  Add tests.